feat: improve structure
