+++ packages/nbt/examples/hello-world.js
@@ -1,5 +1,5 @@
-import { parseNBT } from '../dist/index.js';
 import { readFile } from 'node:fs/promises';
+import { parseNBT } from '../dist/index.js';
 
 const file = await readFile('./examples/servers.dat');
 console.log(parseNBT(file));
