+++ packages/nbt/src/parser/utils.ts
@@ -1,0 +1,3 @@
+export interface ParseReturnOptions {
+  currentOffset: number;
+}
