+++ packages/nbt/tsup.config.ts
@@ -2,5 +2,4 @@ import { createTsupConfig } from '../../tsup.config.js';
 
 export default createTsupConfig({
   entry: ['src/index.ts'],
-  dts: false
 });
